requirements.txt
aioquic
asgiref
dnslib
httpbin
starlette
wsproto
pillow
numpy
